[Chore] Separate out `vllm.utils.network_utils` (#27164)
