+++ tests/utils.py
@@ -46,9 +46,9 @@ from vllm.platforms import current_platform
 from vllm.transformers_utils.tokenizer import get_tokenizer
 from vllm.utils import (
     FlexibleArgumentParser,
-    get_open_port,
 )
 from vllm.utils.mem_constants import GB_bytes
+from vllm.utils.network_utils import get_open_port
 from vllm.utils.torch_utils import cuda_device_count_stateless
 
 if current_platform.is_rocm():
