+++ vllm/config/parallel.py
@@ -18,7 +18,7 @@ from vllm.model_executor.layers.batch_invariant import (
     vllm_is_batch_invariant,
 )
 from vllm.platforms import current_platform
-from vllm.utils import get_open_ports_list
+from vllm.utils.network_utils import get_open_ports_list
 from vllm.utils.torch_utils import cuda_device_count_stateless
 
 if TYPE_CHECKING:
