+++ vllm/entrypoints/openai/api_server.py
@@ -115,9 +115,9 @@ from vllm.utils import (
     Device,
     FlexibleArgumentParser,
     decorate_logs,
-    is_valid_ipv6_address,
     set_ulimit,
 )
+from vllm.utils.network_utils import is_valid_ipv6_address
 from vllm.v1.engine.exceptions import EngineDeadError
 from vllm.v1.metrics.prometheus import get_prometheus_registry
 from vllm.version import __version__ as VLLM_VERSION
